eventheader_types - new crate factored from eventheader
Create a new `eventheader_types` crate. This contains type definitions
for eventheader data structures that are used by both producers
(Linux-only) and consumers (cross-platform) of eventheader-encoded data.

- Version number for `eventheader` and `eventheader_dynamic` incremented
  to `v0.4.1`.
- `eventheader` crate now depends on `eventheader_types`.
- Move "EventHeader Technical Details" docs to `eventheader_types`.
- Move `descriptors`, `enums`, and `guid` modules to
  `eventheader_types` (except for `slice_count`, which remains in
  `eventheader`).
