+++ eventheader_types/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "eventheader_dynamic"
+name = "eventheader_types"
 version = "0.4.1"
 edition = "2021"
 authors = ["Microsoft"]
 license = "MIT"
-description = "Rust API for runtime-specified eventheader-encoded Linux Tracepoints via user_events"
+description = "Rust types for eventheader-encoded Linux Tracepoints via user_events"
 keywords = [
     "user_events",
     "eventheader",
@@ -23,12 +23,7 @@ repository = "https://github.com/microsoft/LinuxTracepoints-Rust"
 readme = "README.md"
 rust-version = "1.63"
 
-[features]
-default = ["user_events"]
-user_events = ["eventheader/user_events"] # Logging is enabled if Linux && user_events.
-
 [dependencies]
-eventheader = { default-features = false, version = "= 0.4.1", path = "../eventheader" }
 
 [dev-dependencies]
 uuid  = ">= 1.1"
