+++ eventheader_types/src/changelog.rs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+//! Release history
+
+#[allow(unused_imports)]
+use crate::*; // For docs
+
+/// # v0.4.1 (TBD)
+/// - Move eventheader types from `eventheader` into new `eventheader_types` crate.
+pub mod v0_4_1 {}
